Abstract cache keys to helper module
